refactor(guidelines): apply review suggestions (renaming and moving files)
