+++ src/components/catalog/catalog-tab.element.ts
@@ -1,7 +1,7 @@
 import { html, LitElement, TemplateResult } from 'lit'
 import { customElement } from 'lit/decorators'
 
-import './catalog-element'
+import './catalog.element'
 import '../theme-selector/theme-selector.element'
 
 @customElement('lux-catalog-tab')
